refactor: refactor Controller and DAO classes
